streamlit/requirements.txt
streamlit==0.86.0
mlflow==1.19.0
pandas==1.3.2
matplotlib==3.4.3
tensorflow==2.5.0
Pillow==8.3.1
altair==4.1.0
numpy==1.18.5
